[BE] refactor: fixture 패키지 생성 #169
