+++ backend/src/test/java/com/onair/hearit/auth/infrastructure/client/KakaoUserInfoClientTest.java
@@ -1,6 +1,6 @@
 package com.onair.hearit.auth.infrastructure.client;
 
-import com.onair.hearit.IntegrationTest;
+import com.onair.hearit.fixture.IntegrationTest;
 import com.onair.hearit.auth.dto.response.KakaoUserInfoResponse;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Disabled;
