+++ backend/src/test/java/com/onair/hearit/auth/presentation/AuthKakaoLoginControllerTest.java
@@ -6,7 +6,7 @@ import static io.restassured.RestAssured.given;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.client.WireMock;
-import com.onair.hearit.IntegrationTest;
+import com.onair.hearit.fixture.IntegrationTest;
 import com.onair.hearit.auth.dto.request.KakaoLoginRequest;
 import com.onair.hearit.auth.dto.response.TokenResponse;
 import com.onair.hearit.auth.infrastructure.client.KakaoUserInfoClient;
